+++ src/hooks/tattoo/main.mjs
@@ -200,6 +200,7 @@ function addCharacterSheetContext(doc, buttons) {
  */
 export function addTattooConsumable() {
   if (game.modules.get("dnd-tashas-cauldron")?.active) return;
+  if (game.settings.get("ddb-importer", "disable-tattoo-type")) return;
   CONFIG.DND5E.consumableTypes["tattoo"] = {
     label: "Spellwrought Tattoo",
   };